change semantics of 'beside' (hence also (|||) and (===)) so the result's origin is the same as that of the first argument. This has several advantages:
  1. beside v, (|||), and (===) are now associative
  2. getting the old behavior from the new is easy (just insert an extra align);
     getting the new behavior from old was difficult.
  3. fits better with behavior of cat.

Of course, this may break diagrams which happened to rely on the old
behavior (namely, to put the local origin of the combined diagram at
the point of tangency between the two arguments).  But this is easy to
fix, either by aligning the first argument in the same direction
before using 'beside', or, if the arguments are the same width,
centering the result.