D.Combinators: implement cat' using a balanced folding scheme (fixes #57)
Thanks to Felipe Lessa for the idea.  The usual linear fold was O(n^2)
because appending each new diagram to the accumulated list has to
query its bounding function, which is built up as the iterated maximum
of the O(n) bounding functions of the component diagrams.  Doing a
balanced fold is thus only O(n log n) (and seems to give a big speedup
in practice).  There is also probably more room for optimizing
computation of bounding functions (using memoization and/or retaining
a bounding box for the common special cases of querying in a direction
parallel to an axis) but that can be left to future work.

The new implementation is also considerably prettier and shorter than
the old, so it's a win all around. =)