generalize fill color to support 'recommended' and 'committed' color
Based on the new Data.Monoid.Recommend module from the monoid-extras
package.  The idea is that a "committed" fill color always overrides a
"recommended" color; otherwise (for two committed colors or two
recommended colors) the inner overrides the outer as usual.  The
motivation for this change is supporting text in the SVG backend,
which must explicitly set the fill color to transparent and hence was
not showing text with no fill color explicitly set.  The solution is
now to make the 'text' function (and related functions) set the
recommended fill color to black.  Then users can still override it
with calls to 'fc', but the SVG backend can know to draw the text
with a fill color of black by default.  The only change to the SVG
backend is that it must now set a *recommended* fill color of
transparent for all diagrams, to avoid overriding the recommended
setting for text.