bump upper bound to allow NumInstances-1.2
Note this change is *not* cherry-picked from master, where in fact we
depend on NumInstances == 1.2.* and import only
Data.NumInstances.Tuple.  However, that means that we have removed
some exported instances (the Num, Eq, and Show instances for function
types) and hence have to release that change with a major version
bump.  So here on the 0.6 maintenance branch we just bump the upper
bound so the old API is maintained but using a newer version of
NumInstances.