change Codomain of Located to Point (Codomain a)
switch from

  Codomain (Located a) = Located (Codomain a)

to

  Codomain (Located a) = Point (Codomain a)

When calling atParam on e.g. a Located (Trail R2), I expect to get a
P2, not a Located R2 (though of course the latter can be converted to
the former).

This also means that instead of requiring constraints of the form

  V (Codomain a) ~ V a

we instead have

  Codomain a ~ V a

which makes more sense: we can only add a location to
translation-invariant things, and translation-invariant parametric
things are precisely those things whose codomain is a vector space.