Tied the offsetSegment epsilon to radius.
The default epsilon was much to small for typical uses of offset.  But
it isn't clear what concrete epsilon to use.  Now epsilon is relative to
the given radius.  You can read it as a parameter that lets you get an
offset "within epsilon percent" of the true offset.

This does not quite fix #118.