D.TwoD.Arrow: Use existing line color as default for tail, head, and shaft
This is another cool thing enabled by doing arrows as delayed leaves:
in addition to knowing the overall transformation in effect when
drawing an arrow, we know the overall *style*!  So we can grab
e.g. the line color and use it as a fill color for the arrow head and
tail.  In particular, this patch sets things up so that line color is
used by default for the head, tail, and shaft of an arrow.  That is,
doing (arrow ... # lc blue) will result in a completely blue
arrow (including head and tail).  This can be overridden by setting
headStyle, tailStyle, or shaftStyle as usual.