D.TwoD.Arrow: bug fix: actually take unfrozen transform into account!
Previously, arrow' simply took the tranformation and applied it to the
points origin and (len,0), and then drew an arrow between the images
of those points.  However, the transformation might also be doing some
flipping or shearing which was getting ignored: affine transformations
are not determined by their action on two points.

This was causing, e.g. shafts to be drawn flipped along the axis of
the arrow with the SVG backend, since the SVG backend does a `reflectY`
as the last thing before freezing, but it was getting ignored.

This patch makes `arrow'` properly apply the transformation to the
shaft trail prior to uniformly scaling and positioning it.