D.Attributes: Setters for Style -> {line,fill} color
Note these can't be lenses because there's nothing to return if the
style does not contain a line or fill color attribute. (One could just
return a default value but that would violate the lens laws.)

Seems to work pretty well!  Perhaps we should add setters for other
attributes as well.