Prelude: stop exporting Diagrams.BoundingBox
It exports names 'union' and 'intersection' which clash with names in
the Haskell Prelude; also, when I removed it, nothing in the entire
website etc. broke (with the exception of -contrib:Diagrams.Lens, but
that doesn't count), which is in and of itself a good argument for
removal.  Users can explicitly import it (perhaps qualified) if they
want it.