A potential solution to diagrams/diagrams-cairo#50.
Complicated by the fact that `defaultAnimMainRender` doesn't know what
data type a backend is using for its options, or even whether it has
looping options at all, so it has to take an extra argument which is
`Maybe` a setter for the looping options.  It uses this to set looping
to `False` for all but the last frame.