Remove non-lenses
`_theta` and `_phi` are not Lenses, and cannot be, since they must always
return an Angle in the range [0,1) turn.  `polar`, `cylindrical`, and
`spherical` are removed for the same reason.

`_theta` is kept as a Getter mostly because it is used in this package.
It is marginally more useful as a Getter than as a plain function
because it composes with `rad`, `turn`, and `deg`.