Use default field naming rules for TH-generated lenses
Remove special handling for Proxy fields.
No lenses are made for field names which do not begin with an _.
Name the Proxy field to take advantage of this convention.
The proxy record accessor is not exported, so renaming is not a problem.
Removing this code allows compiling with lens-4.4 as well as earlier.