Fix SegTree section
We fix this by introducing splitAtParam' which results in the split
halves as well as mappings from the original parameter space, to the
space of the split SegTrees.  This mapping can be used by section to
find the right place to make the second split.  Some work still needs to
be done to make splitAtParam' behave sensibly for parameters outside the
range and we should decide if we should include splitAtParam' in the
class as it could be useful in general.  Finally, we should consider if
we want to bake the mapping into the SegTree data structure to make
the split parameter space behave linearly.