bug fix: get line width for arrow shaft
The line width was still Measured in the style, so simply
applying 'getAttr' was always returning Nothing and a default
line width was being used, so setting the line width had no
effect on arrow shafts.  The fix is to use unmeasureAttrs to
first resolve Measured attributes before calling getAttr.