arrows: use global line width for shaft style
Previously it was only taking the line color from the global style.
Now we simply apply the entire global style to the shaft style.
Anything explicitly set in the shaft style will take precedence,
but this means the shaft and joint now pick up not only the ambient
line color but also the line width, so you can do things like

  arrow ... # lw thick # lc red

and it will "just work".