Add Skinned class, make 3D prims useful
Make it easier to manipulate the 3D primitive solids Ellipsoid, Box,
Frustum directly, without converting them to a Diagram.  This is useful
for CSG.

In particular, provide Traced, Enveloped instances, and define the
Skinned instances in terms of these.

The utility functions (sphere, cylinder, &c) now return the primitive
types, not Diagram.  Add the Skinned class to convert these to Diagram,
analogous to `stroke` in 2D.