Added new constructors to font weight
Added constructors to FontWeight data type corresponding to the CSS common weight name mapping. That is, instead of allowing for a user function that accepts a number and reads it as a font weight each constructor represents each of the specified fontweights (100 to 900) with bolder and lighter marking the weight up by 100 points.