fix for: section producing incorrect results #322
The section instance of SegTree seems to be correct, but the
instances for Located, Trail, and Trail' Line all used the
default definition which uses splitAtParam instead of SegTree's
implementation of section. The default definition of splitAtParam
is only correct for types which consist of a single segment or are
parametrized by arclength (rather than with each segment having
equal weight).