bug fix: #329 section on trail producing incorr...
section on trail producing incorrect result when upper param is
>= 1 (domainUpper) #329

The first parameter was not being rescaled properly in the case that
the second parameter was >= 1. The old code tried to rescale as though
an extra segment was added, while what the code actually does is replace
the last segment with a longer one. This leads two two cases, either the
first param is on a previous segment and doesn't need to change or it's
on the last segment and has to be rescaled to the new length.