refactor splitAtParam'. fix section x a b w/ a > b
The three cases on p of splitAtParam' on close inspection only differed
by what is now splitParam, so could be merged. In addition the new
version avoids creating segments of length 0 when p is 0 or 1. I also
decided to remove the dead code for creating the rescale function for
the right segment.

section previously did not handle the case where a > b. I have included
it in this commit because it depends on the splitAtParam' which has
changed signature.